backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.1
pydantic>=2.7.0
pyjwt>=2.8.0
joblib>=1.3.2
xgboost>=2.0.3
scikit-learn>=1.4.1.post1
pymysql>=1.1.0
aiomysql>=0.2.0
aiosqlite>=0.20.0
python-dotenv>=1.0.1
cryptography>=42.0.5
pandas>=2.0.0
